# images/admin_api/src/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"version": "1.0.0",
		"title": "OSI4IOT DEMO Platform",
		"description": "Open source integration for internet of things",
		"license": {
			"name": "Apache 2.0",
			"url": "https://opensource.org/licenses/Apache-2.0"
		}
	},
	"servers": [
		{
			"url": "https://localhost/admin_api/",
			"description": "Url behind proxy"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"security": [
		{
			"bearerAuth": []
		}
	],
...
	"consumes": ["application/json"],
	"produces": ["application/json"],
	"paths": {
...
		"/group_members/user_managed": {		
			"get": {
				"tags": ["Group"],
				"summary": "Get all groups members for groups managed by the logged user",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/GroupMembers"
								}
							}
						}
					},
					"401": {
						"description": "No auth token or invalid token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/Credentials_rejected"
								}
							}
						}
					}
				}
			}
		},				
...
		"/group/{orgId}/{propName}/{propValue}": {
			"parameters": [
				{
					"name": "orgId",
					"in": "path",
					"required": true,
					"description": "Organization id",
					"type": "number"
				},
				{
					"name": "propName",
					"in": "path",
					"required": true,
					"description": "id, name or acronym of the group",
					"type": "string"
				},
				{
					"name": "propValue",
					"in": "path",
					"required": true,
					"description": "Group property value",
					"type": "string"
				}				
			],
			"get": {
				"tags": ["Group"],
				"summary": "Get a group of the organization - Role required: Organization_Admin",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/GroupOutput"
								}
							}
						}
					},
					"400": {
						"description": "Invalid propName",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/InvalidPropName"
								}
							}
						}
					},
					"401": {
						"description": "No auth token or invalid token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/Credentials_rejected"
								}
							}
						}
					},
					"404": {
						"description": "Organization or Group not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/NotFoundMessage"
								}
							}
						}
					}                     
				}
			},
			"patch": {
				"tags": ["Group"],
				"summary": "Update a group of the organization - Role required: Organization_Admin",
				"requestBody": {
					"description": "Group data",
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/definitions/GroupUpdateInput"
							}
						}
					}
				},				
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/Message"
								}
							}
						}
					},
					"400": {
						"description": "Invalid propName",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/InvalidPropName"
								}
							}
						}
					},
					"401": {
						"description": "No auth token or invalid token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/Credentials_rejected"
								}
							}
						}
					},
					"404": {
						"description": "Organization or Group not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/NotFoundMessage"
								}
							}
						}
					}
				}
			},			
			"delete": {
				"tags": ["Group"],
				"summary": "Delete a group of the organization - Role required: Organization_Admin",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/Message"
								}
							}
						}
					},
					"400": {
						"description": "Invalid propName",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/InvalidPropName"
								}
							}
						}
					},
					"401": {
						"description": "No auth token or invalid token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/Credentials_rejected"
								}
							}
						}
					},
					"404": {
						"description": "Organization or Group not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/definitions/NotFoundMessage"
								}
							}
						}
					}                    
				}
			}
		},
...
	},
	"definitions": {
...
		"Credentials_rejected": {
			"type": "object",
			"properties": {
				"message": {
					"type": "string",
					"default": "invalid token"
				},
				"status": {
					"type": "number",
					"default": 401
				}
			}
		},
...
		"Message": {
			"type": "object",
			"properties": {
				"message": {
					"type": "string"
				}
			}
		},
...
		"InvalidPropName": {
			"type": "object",
			"properties": {
				"message": {
					"type": "string",
					"default": "The propName: <propName> is invalid"
				},
				"status": {
					"type": "number",
					"default": 400
				}
			}
		},
		"NotFoundMessage": {
			"type": "object",
			"properties": {
				"message": {
					"type": "string",
					"default": "The entity with <propName>: <propValue> is not found"
				},
				"status": {
					"type": "number",
					"default": 404
				}
			}
		},
...
		"GroupUpdateInput": {
			"properties": {
				"name": {
					"type": "string"
				},
				"acronym": {
					"type": "string"
				},
				"telegramInvitationLink": {
					"type": "string"
				},
				"telegramChatId": {
					"type": "string"
				},				
				"folderPermission": {
					"type": "string"
				}			
			}
		},
...
		"GroupMemberOutput": {
			"properties": {
				"groupId": {
					"type": "string"
				},
				"userId": {
					"type": "string"
				},				
				"firstName": {
					"type": "string"
				},
				"surname": {
					"type": "string"
				},
				"login": {
					"type": "string"
				},							
				"email": {
					"type": "string"
				},
				"telegramId": {
					"type": "string",
					"default": "Viewer"
				},				
				"roleInGroup": {
					"type": "string",
					"default": "Viewer"
				}				
			}
		},			    
		"GroupOutput": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"acronym": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"groupHash": {
					"type": "string"
				},
				"tableHash": {
					"type": "string"
				}
			}
		},
...
		"GroupMembers": {
			"type": "array",
			"items": {
				"$ref": "#/definitions/GroupMemberOutput"
			}
		},
...
	}
}